+++ person/src/homeworks/person/core/service/DancerAction.java
@@ -1,4 +1,4 @@
-package homeworks.person.core.impl;
+package homeworks.person.core.service;
 
 public interface DancerAction extends PersonAction {
     void dancing();
