+++ person/src/homeworks/person/core/service/PersonAction.java
@@ -1,4 +1,4 @@
-package homeworks.person.core.impl;
+package homeworks.person.core.service;
 
 public interface PersonAction {
     void learn();
