+++ person/src/homeworks/person/core/service/PersonFactory.java
@@ -1,4 +1,4 @@
-package homeworks.person.core.impl;
+package homeworks.person.core.service;
 
 public interface  PersonFactory extends PersonAction {
     PersonAction getPersonAction(String personType);
