+++ person/src/homeworks/person/core/service/ProgrammerAction.java
@@ -1,4 +1,4 @@
-package homeworks.person.core.impl;
+package homeworks.person.core.service;
 
 public interface ProgrammerAction extends PersonAction {
     void coding();
