+++ person/src/homeworks/person/core/service/SingerAction.java
@@ -1,4 +1,4 @@
-package homeworks.person.core.impl;
+package homeworks.person.core.service;
 
 public interface SingerAction extends PersonAction {
     void singing();
